Refactor: adição da estrutura feature-based na implementação da home
