+++ frontend/atendimento-app/src/features/home/types.ts
